refactor(ui): split sidebar component files
