+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "repo",
-    "version": "0.5.2",
+    "version": "0.5.3",
     "private": true,
     "scripts": {
         "dev": "turbo dev --filter=!example-* --cache=local:rw",
