+++ packages/ui-react/package.json
@@ -1,6 +1,6 @@
 {
     "name": "twistail-react",
-    "version": "0.5.2",
+    "version": "0.5.3",
     "type": "module",
     "description": "Twistail UI Components",
     "author": "Aris Ripandi <aris@duck.com>",
