+++ packages/ui-react/src/components/sidebar/index.ts
@@ -1,2 +1,6 @@
-export * from './sidebar'
 export * from './sidebar.css'
+export * from './sidebar-utils'
+export * from './sidebar-context'
+export * from './sidebar-main'
+export * from './sidebar-layout'
+export * from './sidebar-menu'
