+++ packages/ui-react/src/components/sidebar/sidebar-utils.tsx
@@ -1,0 +1,28 @@
+import * as React from 'react'
+
+/**
+ * Hook to detect if the current viewport is mobile based on breakpoint
+ */
+export function useIsMobile(breakpoint = 768) {
+  const [isMobile, setIsMobile] = React.useState<boolean | undefined>(undefined)
+
+  React.useEffect(() => {
+    const mql = window.matchMedia(`(max-width: ${breakpoint - 1}px)`)
+    const onChange = () => {
+      setIsMobile(window.innerWidth < breakpoint)
+    }
+    mql.addEventListener('change', onChange)
+    setIsMobile(window.innerWidth < breakpoint)
+    return () => mql.removeEventListener('change', onChange)
+  }, [breakpoint])
+
+  return !!isMobile
+}
+
+// Constants
+export const SIDEBAR_COOKIE_NAME = 'sidebar_state'
+export const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7
+export const SIDEBAR_WIDTH = '16rem'
+export const SIDEBAR_WIDTH_MOBILE = '18rem'
+export const SIDEBAR_WIDTH_ICON = '3rem'
+export const SIDEBAR_KEYBOARD_SHORTCUT = 'b'
