+++ packages/utils/package.json
@@ -1,6 +1,6 @@
 {
     "name": "twistail-utils",
-    "version": "0.5.2",
+    "version": "0.5.3",
     "type": "module",
     "license": "MIT",
     "description": "A set of utility functions used by Twistail",
