+++ website/package.json
@@ -1,6 +1,6 @@
 {
     "name": "website",
-    "version": "0.5.2",
+    "version": "0.5.3",
     "type": "module",
     "private": true,
     "scripts": {
