+++ website/src/content/changelog/0.5.3.mdx
@@ -1,0 +1,8 @@
+---
+title: v0.5.3 (2025/03/27)
+---
+
+- Changes: split [Sidebar](/docs/components/sidebar) component files.
+
+**Contributors:**
+[@riipandi](https://github.com/riipandi)
